+++ Mailer/extensions/mailer-module/src/Api/v1/Handlers/Users/EmailChangedHandler.php
@@ -3,22 +3,18 @@ declare(strict_types=1);
 
 namespace Remp\MailerModule\Api\v1\Handlers\Users;
 
-use Remp\MailerModule\Repositories\UserSubscriptionsRepository;
+use Remp\MailerModule\Models\Users\UserManager;
 use Tomaj\NetteApi\Handlers\BaseHandler;
-use Tomaj\NetteApi\Params\InputParam;
 use Tomaj\NetteApi\Params\PostInputParam;
 use Tomaj\NetteApi\Response\JsonApiResponse;
 use Tomaj\NetteApi\Response\ResponseInterface;
 
 class EmailChangedHandler extends BaseHandler
 {
-    private $userSubscriptionsRepository;
-
     public function __construct(
-        UserSubscriptionsRepository $userSubscriptionsRepository
+        private UserManager $userManager,
     ) {
         parent::__construct();
-        $this->userSubscriptionsRepository = $userSubscriptionsRepository;
     }
 
     public function params(): array
@@ -34,17 +30,12 @@ class EmailChangedHandler extends BaseHandler
         $originalEmail = $params['original_email'];
         $newEmail = $params['new_email'];
 
-        $subscriptions = $this->userSubscriptionsRepository->findByEmail($originalEmail);
-
-        if (empty($subscriptions)) {
+        $result = $this->userManager->changeEmail($originalEmail, $newEmail);
+        if (!$result) {
             return new JsonApiResponse(
                 404,
-                ['status' => 'error', 'message' => 'No user subscriptions for email: ' . $originalEmail]
+                ['status' => 'error', 'code' => 'no_subscription_found', 'message' => 'No user subscriptions for email: ' . $originalEmail]
             );
-        }
-
-        foreach ($subscriptions as $subscription) {
-            $this->userSubscriptionsRepository->update($subscription, ['user_email' => $newEmail]);
         }
 
         return new JsonApiResponse(200, ['status' => 'ok']);