+++ Mailer/extensions/mailer-module/src/Events/BeforeUserEmailChangeEvent.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Remp\MailerModule\Events;
+
+class BeforeUserEmailChangeEvent
+{
+    public function __construct(
+        public readonly string $originalEmail,
+        public readonly string $newEmail,
+    ) {
+    }
+}